requirements.txt
pandas
networkx
scipy
matplotlib
numpy
pyyaml
torch
scikit-learn
openpyxl

# dashboard & explainability
streamlit>=1.32.0
plotly>=5.15.0
shap>=0.44.0

# deep learning & RL
torch>=2.0.0
stable-baselines3>=2.2.0
gymnasium>=0.29.0

# graph
torch-geometric>=2.4.0

# stage 12
xgboost>=2.0.0
faiss-cpu>=1.7.0
